% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sba}
\alias{sba}
\title{Artificially generated data alike that of Suspension bead arrays}
\format{A list that consists of \code{"plate"} which is a \code{factor} of
        plate number, \code{"X"} that contains measured values where columns 
        are targets and rows are samples (or observations).}
\usage{
data(sba)
}
\description{
The data that has similarity to Suspension bead arrays data.
}
\examples{
data(sba)

# plot to check difference of geometric mean of every target between plates
sba_gm <- by(sba$X, sba$plate, apply, 2, function(x) exp(mean(log(x))))	
par(mfrow= c(2, 3))
apply(combn(4, 2), 2, function(ea) {
	plot(sba_gm[[ea[1]]], sba_gm[[ea[2]]], xlab= names(sba_gm)[ea[1]], 
	     ylab= names(sba_gm)[ea[2]], log= "xy", asp= 1)
	abline(0, 1, col= "cadetblue")
})

# show first 10 observations in plate 1 and plate 2
print(sba$X[c(1:10, 97:106), 1:10])		

}
\keyword{datasets}

